floating point types (#108)
* change overlap matrix in compareutils.py to float32

* removed unneeded astype in wcsutil.py

* remove unneeded astype in layer.py

* first attempt to connect to pyimcom_interface interpolations

* ruff reformat of imdestripe.py

* 2nd round ruff reformat of imdestripe.py

* convert arrays in imdestripe.py to use_output_float

* fixed missing colon in imdestripe.py, other formatting

* ruff reformat

* Additional tests notparallel (#110)

* added a test for imdestripe that doesn't execute in another thread

* added a test for the Eigen kernel

* forkserver fix, v1

* added multikappa test

* added iterative multi test

* removed undesired additions in tests

* consolidated 1 line in layer_diagnostics.py

This was a simple integer operation that could be done more elegantly this way.

* added test for splitpsf main (legacy)

---------


* changed paralellization, now 25% faster

* multithreaded overlap computation

* removed broken multithreading option for now

* added integrated test

* one more test

* added test for missing imports

* added test cleanup to save disk space

* added options for faster/less accurate in PyIMCOM_WCS

* Update docstrings in imdestripe.py

---------

+++ src/pyimcom/layer.py
@@ -811,7 +811,7 @@ class CplxNoise:
 
         """
 
-        this_array = np.zeros((4096, 4096))
+        this_array = np.zeros((4096, 4096), dtype=np.float32)
         rng = np.random.default_rng(seed)
         len_ = 8192 * 128
 
@@ -837,7 +837,7 @@ class CplxNoise:
             else:
                 this_array[:, xmin:xmax] = block.reshape((4096, 128))[:, ::-1]
 
-        return this_array[4:4092, 4:4092].astype(np.float32)
+        return this_array[4:4092, 4:4092]
 
 
 class Mask:
